+++ test/login.ts
@@ -1,0 +1,9 @@
+import { ADTClient } from "../src"
+
+export function create() {
+  return new ADTClient(
+    process.env.ADT_URL!,
+    process.env.ADT_USER!,
+    process.env.ADT_PASS!
+  )
+}