+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.2.4] - 2019-01-21
+
+### Added
+
+- tslint enabled
+- transport selection and creation for an object
+- reentrance ticket
+
 ## [0.2.3] - 2019-01-20
 
 ### Added